New helper class SAPdmManager
