+++ SolidWorksAutomator/MacrosProperty/SetAuthorMacro.cs
@@ -48,9 +48,7 @@ namespace SolidWorksAutomator
             // Get user from PDM
             IEdmVault5 valut = new EdmVault5();
 
-            string userName = "";
-
-            userName = GetPdmUserName(valut);
+            string userName = SAPdmManager.GetPdmUserName(valut);
 
             // Get the list of selected components
             List<Component2> vComp = new List<Component2>();
@@ -81,32 +79,6 @@ namespace SolidWorksAutomator
 
                 prpManager.SetCustomProperty(models[i], GlobalConfig.AuthorPropName, userName);                
             }
-        }
-
-        /// <summary>
-        /// Get the username from PDM
-        /// </summary>
-        /// <param name="pdmValut">The istance to the Vault</param>
-        /// <returns>The string of the username connected to the Vault</returns>
-        public static string GetPdmUserName(IEdmVault5 pdmValut)
-        {
-            pdmValut.LoginAuto(GlobalConfig.VaultName, 0);
-
-            if (pdmValut.IsLoggedIn)
-            {
-
-                IEdmUserMgr5 userMgr = (IEdmUserMgr5)pdmValut;
-
-                IEdmUser6 user = (IEdmUser6)userMgr.GetLoggedInUser();
-
-                return (string)user.UserData;
-
-            }
-            else
-            {
-                return GlobalConfig.AuthorDefault;
-            }
-
         }
     }
 }
